requirements.txt
asgiref==3.7.2
dataclasses==0.6
distlib==0.3.8
Django==4.1.13
django-mongodb-engine @ git+https://github.com/django-nonrel/mongodb-engine@1314786975d3d17fb1a2e870dfba004158a5e431
djangorestframework==3.14.0
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.6
dnspython==2.4.2
filelock==3.13.1
mongoengine==0.27.0
#paho-mqtt==1.6.1
platformdirs==4.1.0
pymongo==4.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
setuptools==69.0.3
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
gunicorn==22.0.0
django==4.1.13
paho-mqtt==2.1.0
requests==2.28.1

